requirements.txt
csv
selenium
os
python-dotenv